refactor(sales-by-person): extract models to separate file and improve styling consistency
