+++ apps/web-giddh/src/app/new-vs-old-Invoices/sales-by-person/sales-by-person.component.scss
@@ -6,10 +6,6 @@
         border: 1px solid var(--border-color);
         padding: 8px 12px;
         box-sizing: border-box;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-        vertical-align: middle;
     }
 }
 
@@ -103,32 +99,17 @@
 .sbp-leaf-th {
     font-weight: var(--font-weight-semibold);
     background-color: var(--theme-primary-light);
-
-    &--left {
-        text-align: left !important;
-    }
-
-    &--right {
-        text-align: right !important;
-    }
 }
 
 // ── Data cells ────────────────────────────────────────────
 .sbp-td {
     &--left {
-        text-align: left;
         font-weight: var(--font-weight-medium);
-    }
-
-    &--right {
-        text-align: right;
     }
 }
 
 // ── Expandable data rows ──────────────────────────────────
 .sbp-data-row--expandable {
-    cursor: pointer;
-
     &:hover .mat-mdc-cell {
         filter: brightness(0.95);
     }
@@ -151,12 +132,10 @@
 .sbp-expand-detail-td {
     padding: 12px 16px 16px !important;
     background-color: var(--table-background-color);
-    border-top: none !important;
 }
 
 .sbp-expand-loading {
     padding: 16px;
-    text-align: center;
 }
 
 .sbp-expand-empty {
@@ -195,10 +174,10 @@
     }
 }
 
-// ── Shared grid layout (4 columns: period flex, 3 fixed num) ─
+// ── Shared grid layout (4 columns: period 30%, 3 num columns share remaining 70%) ─
 %sbp-expand-grid {
     display: grid;
-    grid-template-columns: 1fr 120px 160px 120px;
+    grid-template-columns: 30% repeat(3, 1fr);
     align-items: center;
     padding: 0 16px;
 }
@@ -218,7 +197,6 @@
     &__num {
         font-weight: var(--font-weight-semibold);
         letter-spacing: 0.3px;
-        text-align: right;
     }
 }
 
@@ -246,8 +224,6 @@
     }
 
     &__num {
-        text-align: right;
-
         &--link {
             color: var(--theme-primary-color);
             cursor: pointer;
